All data moved to backed, fetching using http requests
