+++ data/projects/website.js
@@ -1,0 +1,20 @@
+module.exports = {
+    title: "Website",
+    icon: "/content/images/logos/website-brand.png",
+    github: "https://github.com/agoddijn/website",
+    descriptionShort: "Portfolio website designed from scratch using MEAN stack technology",
+    sections: [
+      {
+        header: "Introduction"
+      },
+      {
+        header: "Tools"
+      },
+      {
+        header: ""
+      },
+      {
+        header: "Outlook"
+      }
+    ]
+  }
