+++ data/projects/youtube.js
@@ -1,0 +1,72 @@
+module.exports = {
+    title: "YouTube",
+    icon: "/content/images/logos/YouTube.png",
+    video: "https://www.youtube.com/channel/UCU6lCeBsZp-aWJ6_ymRncBw",
+    descriptionShort: "A YouTube channel on which I have posted tutorial videos",
+    sections: [
+      {
+        header: "World of Warcraft Servers",
+        descriptionAbove: "I used to play a lot of World of Warcraft, and I wanted to play the game for free on servers that let you access the games more advanced content faster. I took to YouTube to teach me how to connect to private servers that people hosted on their computers. Once I figure out how to do it, the tutorials methods seemed overcomplicated, so I made my own videos to show how simple it was to do.",
+        descriptionBelow: "After some moderate success, I decided to host my own server. This was quite challenging, but again I found the online tutorials too complicated, and made my own. One problem I had was particularly difficult to get answers for online, so once I had it figured out I made my own, and this one turned out to have the most success.",
+        additional: [
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/AvPJUOEGOVI?list=PL8EF768622370E00D",
+            caption: "How to play WoW for free"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/Xy1rwxuM7UE?list=PL8EF768622370E00D",
+            caption: "How to make a WoW server"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/lCbGKl6cCgA?list=PL8EF768622370E00D",
+            caption: "How to make your server public"
+          }
+        ]
+      },
+      {
+        header: "Java tutorials",
+        descriptionAbove: "I was given the opportunity to choose my own project for a school assignment, and after the success of my WoW videos, and my curiosity about programming, I decided to learn some basic Java and make some beginner video tutorials for it. I was overwhelmed by positive feedback from people who, like me, had found that even the beginner tutorials for programming were inaccessible.",
+        descriptionBelow: "I got a good grade for the project, but more importantly learnt that it's difficult to overcome a problem if you don't think you can do it, or if it seems too daunting. The barrier's just an illusion though. the material is usually not too difficult but people get bogged down in the nomenclature and everything seems confusing. Once you manage to translate the problem into something you;re more familiar with however, it becomes easy.",
+        additional: [
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/mosJduwmCGs?list=PL0D68608568DDD969",
+            caption: "Introduction"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/FBeeJz5ivuQ?list=PL0D68608568DDD969",
+            caption: "Basic Elements I"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/sttOAEDfM58?list=PL0D68608568DDD969",
+            caption: "Basic Elements II"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/ZJJAS8qKbjw?list=PL0D68608568DDD969",
+            caption: "Basic Elements III"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/Wb5XuSm0P9E?list=PL0D68608568DDD969",
+            caption: "Tools I"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/U3YnD-KYtBU?list=PL0D68608568DDD969",
+            caption: "Tools II"
+          },
+          {
+            type: "video",
+            content: "http://www.youtube.com/embed/xmCNT4esvzk?list=PL0D68608568DDD969",
+            caption: "Arrays"
+          }
+        ]
+      }
+    ]
+  }
